Add command line parsing and "binary"
The binary requires linux and java to be installed, but does at least allow you to easily run it without having to type `java -jar myfile.jar ...`
Also updated the uploading code as the newer Canvas files service doesn't need the key sorting it appears and the existing code was breaking it. So we now sort correctly incase the do roll back to using S3 directly.